move function calculateMoveInDistance to transportation.ts
